# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SiteServices {
  plex: string;
  overseerr: string;
  tautulli: string;
  radarr: string;
  sonarr: string;
  sabnzbd: string;
  qbittorrent: string;
}

export interface SiteLinks {
  plex: string;
  overseerr: string;
  tautulli: string;
  radarr: string;
  sonarr: string;
  sabnzbd: string;
  qbittorrent: string;
}

export interface SiteBackground {
  style: string;
}

export interface SiteBranding {
  name: string;
  tagline: string;
  accent_color: string;
}

export interface SiteAccess {
  request_url: string;
  request_label: string;
  access_url: string;
  access_label: string;
}

export interface SiteConfig {
  background: SiteBackground;
  branding: SiteBranding;
  services: SiteServices;
  links: SiteLinks;
  access: SiteAccess;
  captcha_site_key?: string;
}

export interface StreamSession {
  user: string;
  title: string;
  parent_title: string;
  grandparent_title: string;
  media_type: string;
  progress_percent: string;
  state: string;
  player: string;
  duration: number;
  view_offset: number;
}

export interface ActivityData {
  stream_count: number;
  sessions: StreamSession[];
  configured: boolean;
}

export interface PostersData {
  posters: string[];
  source: string;
}

export interface ServiceStatus {
  ok: boolean;
  latency_ms: number;
  configured: boolean;
}

export interface HealthCheckData {
  plex: ServiceStatus;
  overseerr: ServiceStatus;
  tautulli: ServiceStatus;
  radarr: ServiceStatus;
  sonarr: ServiceStatus;
  sabnzbd: ServiceStatus;
  qbittorrent: ServiceStatus;
}

export interface MediaRequest {
  id: number;
  title: string;
  poster_path: string;
  media_type: string;
  request_status: number;
  media_status: number;
  requested_by: string;
}

export interface RequestsData {
  requests: MediaRequest[];
  configured: boolean;
}

export interface DownloadSlot {
  filename: string;
  percentage: string;
  size: string;
  sizeleft: string;
  status: string;
  timeleft: string;
  cat: string;
  source: string;
}

export interface DownloadHistorySlot {
  filename: string;
  size: string;
  source: string;
  completed_at?: number | null;
}

export interface DownloadsData {
  speed: string;
  kbpersec: string;
  mb: string;
  mbleft: string;
  diskspace1: string;
  noofslots: number;
  slots: DownloadSlot[];
  configured: boolean;
  qbt_configured: boolean;
  history: DownloadHistorySlot[];
}

export interface RecentlyAddedItem {
  title: string;
  grandparent_title: string;
  parent_title: string;
  media_type: string;
  thumb_url?: string | null;
  added_at: number;
  year?: number | null;
}

export interface RecentlyAddedData {
  items: RecentlyAddedItem[];
  configured: boolean;
}

export interface CpuMetrics {
  usage_percent: number;
  temp_celsius?: number | null;
  cores: number;
  brand: string;
  load_1m: number;
  load_5m: number;
  load_15m: number;
}

export interface MemoryMetrics {
  total_bytes: number;
  used_bytes: number;
  free_bytes: number;
  usage_percent: number;
}

export interface DiskMetric {
  fs: string;
  mount: string;
  size_bytes: number;
  used_bytes: number;
  use_percent: number;
}

export interface NetworkMetrics {
  rx_bytes_per_sec: number;
  tx_bytes_per_sec: number;
  iface: string;
}

export interface GpuMetric {
  vendor: string;
  model: string;
  vram_mb?: number | null;
  temp_celsius?: number | null;
  usage_percent?: number | null;
  memory_used_mb?: number | null;
  memory_total_mb?: number | null;
  memory_usage_percent?: number | null;
}

export interface ServerMetrics {
  cpu: CpuMetrics;
  memory: MemoryMetrics;
  disks: DiskMetric[];
  network: NetworkMetrics;
  uptime_seconds: number;
  hostname: string;
  gpus: GpuMetric[];
}

export interface AuthStatus {
  authenticated: boolean;
}

export interface AuthCredentials {
  password: string;
}

export interface AuthResponse {
  success: boolean;
  message: string;
}

export interface AccessRequestBody {
  name: string;
  plex_username: string;
  email: string;
  message?: string;
  turnstile_token: string;
  _hp?: string;
  _ts: number;
}

export interface AccessRequestResponse {
  success: boolean;
  message: string;
}